fix(multi-monitor): resolve slot-math display from the space, not NSScreen.main (#449)
Keyboard resize (span, BSP focus sign, scrolling slot), mouse
resize-end, the float nudge's off-screen fallback, and the border
ring's pixel-scale screen all resolved geometry against
NSScreen.main, so on a secondary monitor they computed against the
wrong bounds. Resolve via the space's own display
(TilingEngine.screen(for:in:)) and, for the ring, the max-overlap
screen (screen(containing:)) so a stash-corner frame with an
off-screen center still finds its display.

Splits KiwiCore+Resize into +ResizeBsp / +ResizeStack for the
350-line ceiling. Single-monitor and headless (test) fallbacks are
unchanged: screen(for:in:) still ends at main ?? first.

fixes #449
